+++ defender/datadog_alerts_heat_map/autotask-1/index.js
@@ -117,12 +117,10 @@ function createAlertsQuery(botId, currentTimestamp, lastUpdateTimestamp) {
           }
         }
         nextPageValues {
-          blocknumber
           timestamp
           id
         }
         currentPageValues {
-          blocknumber
           timestamp
           id
         }
